Update browser.Handler OpenFile behaviour for ext
Do not allow file editing on external window handlers.
This also happens to be a workaround around plugins exiting
upon trying to open a file, expecting that the plugin window
is going to close but not closing because OpenFile swaps the
plugin handler before the rpc Handler processes the exit
return from a HandleResponse (see handler/rpc_breaker.go).

Adds handler.WindowManager.Shiftable() which enables clients to query
whether a ShiftFocus call would be successful and to which node the
focus would be switched.