Remove handler.Client rpc breaker logic
When multiple breakers are using global interrupts it's hard to figure
out when to re-issue a draw because the handler.Server is requesting
vs some other handler.Server is requesting it and it causes resource
starvation problems. For now a new short timeout is set to minimize
plugins slowing down the host's runtime.