Fix workspace race via workspace server
Main event loop used a different sync.Locker so workspace requests
were not synchronized properly. This also updates workspace.Executor
impls to cleanup after Wait because that's the semantics of cmd.Exec.Wait.

We need to find a better way without burdening Scheme impls
with synchronization but for now this will suffice.