Fix temporary bluectx.First goroutine leaks
We must ensure that at least one of the ctx passed to bluectx.First
is canceled. This ctx is not exposed via api so in practice it's
set to the workspace rpc server tied to the lifecycle of a plugin
and the tracking workspace, tied to the lifecycle of a workspace.
This resulted in goroutine leaks until the workspace was closed.