Deprecate textapi.Editor.Edit
Public clients should not have direct access to the internal
editor implementations, since there are some side-effects like
having no EventTypeFocus support that make it a confusing API
to use. The only reason why there was such method in the first place
is because internal text.Editor implementations need to
implement Edit for text.Component to make use of them.