Introduce partial support for grapheme clusters
This effectively enables re-processing of individually inserted grapheme
clusters of runes, as well as ReadFrom support. It also adds Width to term.Cell
and this is calculated once upon initial processing and not for every render,
which previously would call runewidth.RuneWidth for every single cell.

Missing full support from tcell: currently writer would need to ensure that
when a multi width grapheme cluster is encounted, all the cells to the right of it
are moved accordingly. This breaks encapsulation of the basic primitives of this library
as tui.Component's should not need to be aware of its surroundings to render
correctly. A possible solution would be to have a caching layer at the term.Writer
level and upon Flush identify any grapheme clusters and shift
runes accordingly. Another approach would be to just refactor tcell to handle
this case more gracefully.