Fix scroll offset + 1
The reason behind this heuristics is unknown but it was
probably related to paste or just ease of cursor navigation.
The problem is that this causes the offset to be changed even
when content is perfectly aligned with height/width.

If padding is needed for x offset, then that should be handled
at a higher level like cursor and use SetOffset to force it.