Update vte vi to improve performance on slow shells
This ensures that we don't apply backpressure
to the user in slow shells (i.e. over ssh) until we reach 50
buffered events, which is quite a lot.

The maximum number of events that tcell can take is 100, so 50
is a safe number which prevents deadlocks.

This also removes a second call to schedule a bell callback
on every handle event on vte's vi handler, which
saves a great deal of latency as well for the user at the expense
of a cursor that's potentially out of bounds in some edge cases.