Refactor term.Event ctrl keys to use ModCtrl
tcell exposes way too many of the implementation details of the vte protocol;
this is a first attempt at reducing exposure.

Also adds support for any arbitrary combination of modifiers, forcing
handlers to not assume that Modifier is not "usually" set. This is
important to enable better key binding support if we were to adopt
the kitty keyboard protocol or implement the GUI.