Update gui's interrupts to reduce extra calls to Draw
Previously mixed payloads in interrupts would cause handler
to Draw twice per call to Update.

We are also now assuming that if an interrupt is received
with a UserFunc, that doesn't translate to a call to Draw.

This improves the perceived performance when running compute
intensive shaders. When measured in TPS the improvement is
around ~10x.