Add intro shader: Burning
Burning shows fire building the given logo as flames touch it.

As flame heat caresses the logo it ignites it and dissipates after some
configurable time. Brighter parts of the logo will take longer to dissipate
heat.

Then flames start fading out and a cyclic ember ripples form, moving from
darker parts of the logo to brighter ones (logo gradient) sampling from a
configurable color gradient, conveying the idea it is radiating out the heat
it has absorbed.

To generate what we call the "logo gradient", which is a numerical value
denoting the brightness of the logo cell, we use the [asciiart.Encode] using
numbers as density characters.

In order to perform effects on the wallpaper asciiart image we use the
following contract: The wallpaper will have its background filled with
special invisible characters. That way we know the bounds of the wallpaper
image. Another asciiart image is recreated (logo gradient) and stretched
over the same boundaries effectively covering the original. That way we can
also mask out the origial logo and paint it out to produce the illusion the
logo has disappeared.

Designed taking https://www.shadertoy.com/view/XsXSWS as reference.

Reviewed-on: https://git.unstable.build/unstablebuild/go-tui/pulls/89
Reviewed-by: Ernest Romero Climent <ernest@unstable.build>
Co-authored-by: ramon <ramon@unstable.build>
Co-committed-by: ramon <ramon@unstable.build>