Fix search list double pushing of matched elements
The experience suffers a bit, as now the search results of a very
large input query are only pushed at the end of the search.

On the flip side, this reduces contention, as we only acquire
the lock once at the end of the search, rather than for every
match.