Fix trapped cursor (can't move up) when gone beyond vertical content bounds
Fixed by adding vi cursor mark when cursor moved to fit vertical content bounds.

Before each MoveUp/Down the cursor is placed at the last
vi.free CursorMark. If you jump beyond content, viHandlerImpl bounds the
cursor to be within content scope (it sets the cursor to coords in range),
but does not place a mark.

That creates a problem if you for instance fall 10 lines beyond the content
rows and then try to MoveUp. The cursor will lose its correct position (last row)
and be placed again out of bound beyond last logical row. The cause is
setCursorAtScroll not placing a mark after correcting the coordinate when fell
out of vertical bounds, no need to do it otherwise.

No need to do it for horizontal bounding because there's no action that calls
setCursorAtScroll on horizontal coords (for that you would use `<num>l` and this
does not jump but step one at a time rightwards.

Reviewed-on: https://git.unstable.build/unstablebuild/go-tui/pulls/121
Reviewed-by: Ernest Romero Climent <ernest@unstable.build>
Co-authored-by: ramon <ramon@unstable.build>
Co-committed-by: ramon <ramon@unstable.build>