Fix gui race condition between update and consume events
If the gui mutex, used to protect the variables shared between
the consumeEvents goroutine and the update goroutine is also
used to protect the root handler, then the root handler might
unlock to wait for I/O, accidentally allowing the consumeEvents
goroutine to append new events before the update goroutine cleans
the events slice, which it does after acquiring the lock back.

The deeper problem lies in having to unlock waiting for I/O for
all the extensions to avoid deadlocks when performing API calls.
We should find a way to asynchronously call Handle in a tui.Handler
rpc client, without complicating the implementation too much. Same
for commands, although in this case it's easier since the return
value doesn't really need to be propagated from extension back to host.