Remove textapi.EventTypeRename
Renames coming from FS eventing are dispatched in two events.
To make this simpler for our clients, we would need to add a property
to an already bloated Event struct, and coalesce the FS events into
one. This is possible but not necessary, as we can dispatch an
EventTypeRemove followed by a EventTypeFlush to mimic the rename
behaviour. The differences between rename and remove and flush
are not relevant for our use case. If the need arises, we can
revisit this.